+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 4.8 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Various fixes for the ``PURE_PYTHON`` version, e.g.
+  make ``Acquired`` an ``str`` (as required by ``Zope``),
+  respect base object's ``__eq__``,
+  (`#48 <https://github.com/zopefoundation/Acquisition/issues/48>`_).
 
 
 4.7 (2020-10-07)